requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
multidict==6.0.4
openai==0.27.8
requests==2.31.0
smmap==5.0.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
